LoggingPolicy.Factory -> HttpLoggingPolicyFactory
