+++ client-runtime/src/main/java/com/microsoft/rest/v2/http/HttpPipelineLogger.java
@@ -19,7 +19,7 @@ public interface HttpPipelineLogger {
 
     /**
      * Log the provided message.
-     * @param logLevel The LogLevel associated with this message.
+     * @param logLevel The HttpLogDetailLevel associated with this message.
      * @param message The message to log.
      * @param formattedArguments A variadic list of arguments that should be formatted into the
      *                           provided message.
